Improvement all over the board. Archicteture fixed
