+++ pages/privacy-policy.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import MarkdownVisualizer from "../components/common/MarkdownVisualizer";
+import MarkdownVisualizer from "../components/MarkdownVisualizer";
 import privacyPolicy from "../components/md/privacyPolicy";
 
 const PrivacyPolicy: React.SFC = () => <MarkdownVisualizer markdown={privacyPolicy} />;
