+++ pages/user-terms.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import MarkdownVisualizer from "../components/common/MarkdownVisualizer";
+import MarkdownVisualizer from "../components/MarkdownVisualizer";
 import userTerms from "../components/md/userTerms";
 
 const UserTerms: React.SFC = () => <MarkdownVisualizer markdown={userTerms} />;
